corporate_partner_access/edxapp_wrapper/backends/course_module_v1.py
"""Backend abstraction for course module from edx-platform."""


from openedx.core.djangoapps.content.course_overviews.models import CourseOverview  # pylint: disable=import-error


def course_overview_backend():
    """Return CourseOverview class."""

    return CourseOverview
