requirements/test.txt
#
# This file is autogenerated by pip-compile with Python 3.11
# by the following command:
#
#    pip-compile --output-file=requirements/test.txt requirements/test.in
#
asgiref==3.9.1
    # via
    #   -r requirements/base.txt
    #   django
cffi==1.17.1
    # via
    #   -r requirements/base.txt
    #   pynacl
click==8.2.1
    # via
    #   -c requirements/constraints.txt
    #   -r requirements/base.txt
    #   edx-django-utils
coverage[toml]==7.9.2
    # via pytest-cov
    # via
    #   -c https://raw.githubusercontent.com/edx/edx-lint/master/edx_lint/files/common_constraints.txt
    #   -r requirements/base.txt
    #   django-crum
    #   django-model-utils
    #   django-waffle
    #   edx-django-utils
django-crum==0.7.9
    # via
    #   -r requirements/base.txt
    #   edx-django-utils
django-model-utils==5.0.0
    # via -r requirements/base.txt
django-waffle==5.0.0
    # via
    #   -r requirements/base.txt
    #   edx-django-utils
edx-django-utils==8.0.0
    # via -r requirements/base.txt
iniconfig==2.1.0
    # via pytest
openedx-atlas==0.7.0
    # via -r requirements/base.txt
packaging==25.0
    # via
    #   -c requirements/constraints.txt
    #   pytest
pbr==6.1.1
    # via
    #   -r requirements/base.txt
    #   stevedore
pillow==11.3.0
    # via -r requirements/base.txt
pluggy==1.6.0
    # via
    #   pytest
    #   pytest-cov
psutil==7.0.0
    # via
    #   -r requirements/base.txt
    #   edx-django-utils
pycparser==2.22
    # via
    #   -r requirements/base.txt
    #   cffi
pygments==2.19.2
    # via pytest
pynacl==1.5.0
    # via
    #   -r requirements/base.txt
    #   edx-django-utils
pytest==8.4.1
    # via
    #   pytest-cov
    #   pytest-django
pytest-cov==6.2.1
    # via -r requirements/test.in
pytest-django==4.11.1
    # via -r requirements/test.in
sqlparse==0.5.3
    # via
    #   -r requirements/base.txt
    #   django
stevedore==5.4.1
    # via
    #   -r requirements/base.txt
    #   edx-django-utils

# The following packages are considered to be unsafe in a requirements file:
# setuptools
